# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API docs for ToC Trader",
        "title": "ToC Trader",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.4.0"
    },
    "basePath": "/trade-bot",
    "paths": {
...
        "/condition": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TradeCondition"
                ],
                "summary": "GetLatestTradeCondition",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/simulate.Result"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TradeCondition"
                ],
                "summary": "ImpoprtTradeCondition",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/simulate.Result"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TradeCondition"
                ],
                "summary": "DeletaAllResultAndCond",
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/data/bid-ask": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data"
                ],
                "summary": "ReceiveBidAsk",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bidask.BidAskProto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/data/streamtick": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data"
                ],
                "summary": "ReceiveStreamTick",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/streamtick.StreamTickProto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "bidask.BidAskProto": {
            "type": "object",
            "properties": {
                "bid_ask": {
                    "$ref": "#/definitions/bidask.BidAskProto_BidAskData"
                },
                "exchange": {
                    "type": "string"
                }
            }
        },
        "bidask.BidAskProto_BidAskData": {
            "type": "object",
            "properties": {
                "ask_price": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "ask_volume": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "bid_price": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "bid_volume": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "code": {
                    "type": "string"
                },
                "date_time": {
                    "type": "string"
                },
                "diff_ask_vol": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "diff_bid_vol": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "simtrade": {
                    "type": "integer"
                },
                "suspend": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "attachment": {},
                "response": {
                    "type": "string"
                }
            }
        },
...
        "simulate.Result": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "cond": {
                    "$ref": "#/definitions/simulationcond.AnalyzeCondition"
                },
                "cond_id": {
                    "type": "integer"
                },
                "forward_balance": {
                    "type": "integer"
                },
                "is_best_forward": {
                    "type": "boolean"
                },
                "is_best_reverse": {
                    "type": "boolean"
                },
                "positive_days": {
                    "type": "integer"
                },
                "reverse_balance": {
                    "type": "integer"
                },
                "total_days": {
                    "type": "integer"
                },
                "total_loss": {
                    "type": "integer"
                },
                "trade_count": {
                    "type": "integer"
                },
                "trade_day": {
                    "type": "string"
                }
            }
        },
        "simulationcond.AnalyzeCondition": {
            "type": "object",
            "properties": {
                "close_change_ratio_high": {
                    "type": "number"
                },
                "close_change_ratio_low": {
                    "type": "number"
                },
                "forward_out_in_ratio": {
                    "type": "number"
                },
                "history_close_count": {
                    "type": "integer"
                },
                "open_change_ratio": {
                    "type": "number"
                },
                "reverse_out_in_ratio": {
                    "type": "number"
                },
                "rsi_high": {
                    "type": "number"
                },
                "rsi_low": {
                    "type": "number"
                },
                "ticks_period_count": {
                    "type": "integer"
                },
                "ticks_period_limit": {
                    "type": "number"
                },
                "ticks_period_threshold": {
                    "type": "number"
                },
                "trim_history_close_count": {
                    "type": "boolean"
                },
                "volume_per_second": {
                    "type": "integer"
                }
            }
        },
        "streamtick.StreamTickProto": {
            "type": "object",
            "properties": {
                "exchange": {
                    "type": "string"
                },
                "tick": {
                    "$ref": "#/definitions/streamtick.StreamTickProto_TickData"
                }
            }
        },
        "streamtick.StreamTickProto_TickData": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "ask_side_total_cnt": {
                    "type": "integer"
                },
                "ask_side_total_vol": {
                    "type": "integer"
                },
                "avg_price": {
                    "type": "number"
                },
                "bid_side_total_cnt": {
                    "type": "integer"
                },
                "bid_side_total_vol": {
                    "type": "integer"
                },
                "chg_type": {
                    "type": "integer"
                },
                "close": {
                    "type": "number"
                },
                "code": {
                    "type": "string"
                },
                "date_time": {
                    "type": "string"
                },
                "high": {
                    "type": "number"
                },
                "low": {
                    "type": "number"
                },
                "open": {
                    "type": "number"
                },
                "pct_chg": {
                    "type": "number"
                },
                "price_chg": {
                    "type": "number"
                },
                "simtrade": {
                    "type": "integer"
                },
                "suspend": {
                    "type": "integer"
                },
                "tick_type": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "number"
                },
                "total_volume": {
                    "type": "integer"
                },
                "volume": {
                    "type": "integer"
                }
            }
        },
...
    }
}